BACKPORT: file: let pick_file() tell caller it's done
Let pick_file() report back that the fd it was passed exceeded the
maximum fd in that fdtable. This allows us to simplify the caller of
this helper because it doesn't need to care anymore whether the passed
in max_fd is excessive. It can rely on pick_file() telling it that it's
past the last valid fd.

Cc: Christoph Hellwig <hch@lst.de>
Cc: Giuseppe Scrivano <gscrivan@redhat.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: linux-fsdevel@vger.kernel.org
Change-Id: I0d037a47612154c68b4452d8aa26d5cebcf56107
(cherry picked from commit f49fd6d3c070d08c4ae9696876c7098320e48dab)
[ kxxt: Resolved minor conflict in fs/file.c ]
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>